renamed config.rs to graphql_operation_type.rs
