+++ tc_core/src/blueprint/for_config/mod.rs
@@ -1,8 +1,8 @@
-pub mod config;
+pub mod graphql_operation_type;
 pub mod group_by;
 pub mod key_values;
 pub mod server;
 
-pub use config::*;
+pub use graphql_operation_type::*;
 pub use key_values::*;
 pub use server::*;
